+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     * **TODO** does this mean minor or major version bump?
 ### Fixed
   * findOne uses `first` not `limit`
+  * Allow explicit auth-skip for reloading sequelize models.
 ### Deprecated
 ### Security